# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Anime Gallery Media Server",
    "description": "\nAnime Gallery Media Server API 覆盖前台图片浏览、后台媒体资料维护、批量上传任务、\n元数据导入、部署设置与系统健康检查。\n\n鉴权模型：\n\n- 浏览器后台会话使用 `agms_admin_session` HttpOnly Cookie。\n- 运维脚本建议使用 `Authorization: Bearer <AGMS_API_KEYS value>`，并只授予任务实际需要的权限。\n- 使用 Cookie 会话调用非安全方法时，还需要提交与 `agms_admin_csrf` Cookie 匹配的 `X-CSRF-Token`。\n- API Key 不依赖 Cookie，因此不需要 CSRF；可在后台独立创建、设置到期时间、调整权限或立即撤销。\n- 旧版本升级后，已有 API Key 默认迁移为全部权限，避免现有自动化任务中断。\n",
    "version": "0.1.5"
  },
  "servers": [
    {
      "url": "/",
      "description": "当前部署域名"
    },
    {
      "url": "/api",
      "description": "API 前缀"
    }
  ],
  "paths": {
...
    "/api/settings": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "获取后台设置",
        "operationId": "read_settings_api_settings_get",
        "security": [
          {
            "OperationsApiKey": []
          },
          {
            "AdminSessionCookie": [],
            "CsrfTokenCookie": [],
            "CsrfTokenHeader": []
          }
        ],
        "parameters": [
          {
            "name": "agms_admin_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agms Admin Session"
            },
            "description": "管理员会话 Cookie。运维脚本建议改用 `Authorization: Bearer <api-key>`。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSettingsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要鉴权，或提交的令牌无效。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required"
                }
              }
            }
          },
          "403": {
            "description": "API Key 权限不足，或 Cookie 会话发起写入请求时缺少匹配的 CSRF 请求头。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "CSRF validation failed"
                }
              }
            }
          }
        },
        "description": "返回管理员偏好、前台首页图片绑定、随机图片 API 默认值、上传队列参数、GitHub 更新检查代理和运维 API Key。\n\n所需 API Key 权限：`settings:manage`。",
        "x-api-key-scopes": [
          "settings:manage"
        ]
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "更新后台设置",
        "operationId": "update_settings_api_settings_put",
        "security": [
          {
            "OperationsApiKey": []
          },
          {
            "AdminSessionCookie": [],
            "CsrfTokenCookie": [],
            "CsrfTokenHeader": []
          }
        ],
        "parameters": [
          {
            "name": "agms_admin_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agms Admin Session"
            },
            "description": "管理员会话 Cookie。运维脚本建议改用 `Authorization: Bearer <api-key>`。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminSettingsUpdate"
              },
              "examples": {
                "uploadWorkers": {
                  "summary": "调整上传 worker、重试与保留期限",
                  "value": {
                    "upload_worker_count": 24,
                    "upload_claim_batch_size": 4,
                    "upload_task_max_attempts": 3,
                    "upload_failed_retention_days": 7,
                    "image_manage_view_mode": "waterfall",
                    "github_proxy_url": "https://gh-proxy.example.com/"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSettingsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要鉴权，或提交的令牌无效。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required"
                }
              }
            }
          },
          "403": {
            "description": "API Key 权限不足，或 Cookie 会话发起写入请求时缺少匹配的 CSRF 请求头。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "CSRF validation failed"
                }
              }
            }
          }
        },
        "description": "更新管理员资料、图片管理显示模式、首页幻灯片、前台背景图、随机图片 API 默认值、受数据库并发上限约束的上传 worker 参数和 GitHub 更新检查代理。\n\n所需 API Key 权限：`settings:manage`。",
        "x-api-key-scopes": [
          "settings:manage"
        ]
      }
    },
...
    "/api/settings/auth-secret/rotate": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "轮换登录密钥",
        "operationId": "rotate_admin_auth_secret_api_settings_auth_secret_rotate_post",
        "security": [
          {
            "OperationsApiKey": []
          },
          {
            "AdminSessionCookie": [],
            "CsrfTokenCookie": [],
            "CsrfTokenHeader": []
          }
        ],
        "parameters": [
          {
            "name": "agms_admin_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agms Admin Session"
            },
            "description": "管理员会话 Cookie。运维脚本建议改用 `Authorization: Bearer <api-key>`。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要鉴权，或提交的令牌无效。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required"
                }
              }
            }
          },
          "403": {
            "description": "API Key 权限不足，或 Cookie 会话发起写入请求时缺少匹配的 CSRF 请求头。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "CSRF validation failed"
                }
              }
            }
          }
        },
        "description": "生成新的 `AGMS_AUTH_SECRET`，写入 `.env`，并撤销当前浏览器会话。\n\n所需 API Key 权限：`settings:manage`。",
        "x-api-key-scopes": [
          "settings:manage"
        ]
      }
    },
    "/api/settings/api-keys/reset": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "重置运维 API Key",
        "operationId": "reset_operations_api_keys_api_settings_api_keys_reset_post",
        "security": [
          {
            "OperationsApiKey": []
          },
          {
            "AdminSessionCookie": [],
            "CsrfTokenCookie": [],
            "CsrfTokenHeader": []
          }
        ],
        "parameters": [
          {
            "name": "agms_admin_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agms Admin Session"
            },
            "description": "管理员会话 Cookie。运维脚本建议改用 `Authorization: Bearer <api-key>`。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSettingsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要鉴权，或提交的令牌无效。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required"
                }
              }
            }
          },
          "403": {
            "description": "API Key 权限不足，或 Cookie 会话发起写入请求时缺少匹配的 CSRF 请求头。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "CSRF validation failed"
                }
              }
            }
          }
        },
        "description": "生成新的默认 `AGMS_API_KEYS`，写入 `.env`，并使旧 API Key 立即失效。\n\n所需 API Key 权限：`api_keys:manage`。",
        "x-api-key-scopes": [
          "api_keys:manage"
        ]
      }
    },
    "/api/settings/api-keys": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "获取 API Key 列表",
        "operationId": "read_operations_api_keys_api_settings_api_keys_get",
        "security": [
          {
            "OperationsApiKey": []
          },
          {
            "AdminSessionCookie": [],
            "CsrfTokenCookie": [],
            "CsrfTokenHeader": []
          }
        ],
        "parameters": [
          {
            "name": "agms_admin_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agms Admin Session"
            },
            "description": "管理员会话 Cookie。运维脚本建议改用 `Authorization: Bearer <api-key>`。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiKeyRead"
                  },
                  "title": "Response Read Operations Api Keys Api Settings Api Keys Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要鉴权，或提交的令牌无效。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required"
                }
              }
            }
          },
          "403": {
            "description": "API Key 权限不足，或 Cookie 会话发起写入请求时缺少匹配的 CSRF 请求头。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "CSRF validation failed"
                }
              }
            }
          }
        },
        "description": "返回完整 API Key、权限、有效期及最近调用信息。仅 API Key 管理权限可调用。\n\n所需 API Key 权限：`api_keys:manage`。",
        "x-api-key-scopes": [
          "api_keys:manage"
        ]
      },
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "创建 API Key",
        "operationId": "add_operations_api_key_api_settings_api_keys_post",
        "security": [
          {
            "OperationsApiKey": []
          },
          {
            "AdminSessionCookie": [],
            "CsrfTokenCookie": [],
            "CsrfTokenHeader": []
          }
        ],
        "parameters": [
          {
            "name": "agms_admin_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agms Admin Session"
            },
            "description": "管理员会话 Cookie。运维脚本建议改用 `Authorization: Bearer <api-key>`。"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要鉴权，或提交的令牌无效。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required"
                }
              }
            }
          },
          "403": {
            "description": "API Key 权限不足，或 Cookie 会话发起写入请求时缺少匹配的 CSRF 请求头。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "CSRF validation failed"
                }
              }
            }
          }
        },
        "description": "创建可独立授权和设置有效期的新 API Key。\n\n所需 API Key 权限：`api_keys:manage`。",
        "x-api-key-scopes": [
          "api_keys:manage"
        ]
      }
    },
...
    "/api/settings/api-keys/{key_id}/rotate": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "刷新 API Key",
        "operationId": "rotate_operations_api_key_api_settings_api_keys__key_id__rotate_post",
        "security": [
          {
            "OperationsApiKey": []
          },
          {
            "AdminSessionCookie": [],
            "CsrfTokenCookie": [],
            "CsrfTokenHeader": []
          }
        ],
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Key Id"
            }
          },
          {
            "name": "agms_admin_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agms Admin Session"
            },
            "description": "管理员会话 Cookie。运维脚本建议改用 `Authorization: Bearer <api-key>`。"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "需要鉴权，或提交的令牌无效。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "Authentication required"
                }
              }
            }
          },
          "403": {
            "description": "API Key 权限不足，或 Cookie 会话发起写入请求时缺少匹配的 CSRF 请求头。",
            "content": {
              "application/json": {
                "example": {
                  "detail": "CSRF validation failed"
                }
              }
            }
          }
        },
        "description": "重新生成指定 API Key 的密钥值，保留名称、权限和有效期，并使旧密钥立即失效。\n\n所需 API Key 权限：`api_keys:manage`。",
        "x-api-key-scopes": [
          "api_keys:manage"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiKeyCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 9,
            "minItems": 1,
            "title": "Scopes"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "scopes"
        ],
        "title": "ApiKeyCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "OperationsApiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "AGMS_API_KEY",
        "description": "后台创建的可分权运维 API Key。权限：`library:read` 媒体库读取；`uploads:manage` 图片上传；`library:write` 资料维护；`library:delete` 永久删除；`system:read` 系统只读；`settings:manage` 系统设置；`updates:read` 更新只读；`updates:run` 执行更新；`api_keys:manage` API Key 管理"
      },
      "AdminSessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "agms_admin_session",
        "description": "浏览器管理员会话 Cookie。"
      },
      "CsrfTokenCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "agms_admin_csrf",
        "description": "管理员会话同时签发的 CSRF Cookie。"
      },
      "CsrfTokenHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "使用 Cookie 会话调用非安全方法时，必须与 CSRF Cookie 一致。"
      }
    }
  },
...
}